espnet3: rename iterator module to epoch_sync_iterator
The module holds exactly one class, EpochSyncIterator, and the vague
name no longer describes the contents: the wrapped source is a callable
or container, not an iterator. Follow the espnet convention of naming a
module after its class (as sequence_iter_factory / chunk_iter_factory
do), and move the tests alongside.
